+++ packages/apply-release-plan/src/index.test.ts
@@ -51,8 +51,11 @@ class FakeReleasePlan {
       ignore: [],
       ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
         onlyUpdatePeerDependentsWhenOutOfRange: false,
-        updateInternalDependents: "out-of-range",
-        useCalculatedVersionForSnapshots: false
+        updateInternalDependents: "out-of-range"
+      },
+      snapshot: {
+        useCalculatedVersion: false,
+        prereleaseTemplate: null
       },
       ...config
     };
@@ -74,6 +77,7 @@ async function testSetup(
   fixtureName: string,
   releasePlan: ReleasePlan,
   config?: Config,
+  snapshot?: string | undefined,
   setupFunc?: (tempDir: string) => Promise<any>
 ) {
   if (!config) {
@@ -86,10 +90,13 @@ async function testSetup(
       baseBranch: "main",
       updateInternalDependencies: "patch",
       ignore: [],
+      snapshot: {
+        useCalculatedVersion: false,
+        prereleaseTemplate: null
+      },
       ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
         onlyUpdatePeerDependentsWhenOutOfRange: false,
-        updateInternalDependents: "out-of-range",
-        useCalculatedVersionForSnapshots: false
+        updateInternalDependents: "out-of-range"
       }
     };
   }
@@ -108,7 +115,8 @@ async function testSetup(
     changedFiles: await applyReleasePlan(
       releasePlan,
       await getPackages(tempDir),
-      config
+      config,
+      snapshot
     ),
     tempDir
   };
@@ -488,8 +496,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -551,8 +562,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -653,6 +667,47 @@ describe("apply release plan", () => {
         version: "1.0.0"
       });
     });
+    it("should use exact versioning when snapshot release is applied, and ignore any range modifiers", async () => {
+      const releasePlan = new FakeReleasePlan(
+        [
+          {
+            id: "some-id",
+            releases: [{ name: "pkg-b", type: "minor" }],
+            summary: "a very useful summary"
+          }
+        ],
+        [
+          {
+            changesets: ["some-id"],
+            name: "pkg-b",
+            newVersion: "1.1.0",
+            oldVersion: "1.0.0",
+            type: "minor"
+          }
+        ]
+      );
+      let { changedFiles } = await testSetup(
+        "simple-project-caret-dep",
+        releasePlan.getReleasePlan(),
+        releasePlan.config,
+        "canary"
+      );
+
+      let pkgPath = changedFiles.find(a =>
+        a.endsWith(`pkg-a${path.sep}package.json`)
+      );
+
+      if (!pkgPath) throw new Error(`could not find an updated package json`);
+      let pkgJSON = await fs.readJSON(pkgPath);
+
+      expect(pkgJSON).toMatchObject({
+        name: "pkg-a",
+        version: "1.1.0",
+        dependencies: {
+          "pkg-b": "1.1.0"
+        }
+      });
+    });
 
     describe("internal dependency bumping", () => {
       describe("updateInternalDependencies set to patch", () => {
@@ -700,8 +755,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -785,8 +843,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -862,8 +923,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -939,8 +1003,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -1019,8 +1086,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -1104,8 +1174,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -1181,8 +1254,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -1258,8 +1334,11 @@ describe("apply release plan", () => {
               ignore: [],
               ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
                 onlyUpdatePeerDependentsWhenOutOfRange: false,
-                updateInternalDependents: "out-of-range",
-                useCalculatedVersionForSnapshots: false
+                updateInternalDependents: "out-of-range"
+              },
+              snapshot: {
+                useCalculatedVersion: false,
+                prereleaseTemplate: null
               }
             }
           );
@@ -1339,8 +1418,11 @@ describe("apply release plan", () => {
             ignore: [],
             ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
               onlyUpdatePeerDependentsWhenOutOfRange: true,
-              updateInternalDependents: "out-of-range",
-              useCalculatedVersionForSnapshots: false
+              updateInternalDependents: "out-of-range"
+            },
+            snapshot: {
+              useCalculatedVersion: false,
+              prereleaseTemplate: null
             }
           }
         );
@@ -1501,8 +1583,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -1606,8 +1691,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -1691,8 +1779,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -1780,8 +1871,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -1883,8 +1977,11 @@ describe("apply release plan", () => {
           ignore: [],
           ___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH: {
             onlyUpdatePeerDependentsWhenOutOfRange: false,
-            updateInternalDependents: "out-of-range",
-            useCalculatedVersionForSnapshots: false
+            updateInternalDependents: "out-of-range"
+          },
+          snapshot: {
+            useCalculatedVersion: false,
+            prereleaseTemplate: null
           }
         }
       );
@@ -2088,6 +2185,7 @@ describe("apply release plan", () => {
         "simple-project",
         releasePlan.getReleasePlan(),
         releasePlan.config,
+        undefined,
         setupFunc
       );
 
@@ -2114,6 +2212,7 @@ describe("apply release plan", () => {
         "simple-project",
         releasePlan.getReleasePlan(),
         { ...releasePlan.config, ignore: ["pkg-a"] },
+        undefined,
         setupFunc
       );
 
@@ -2156,6 +2255,7 @@ describe("apply release plan", () => {
         "simple-project",
         releasePlan.getReleasePlan(),
         releasePlan.config,
+        undefined,
         setupFunc
       );
 
@@ -2212,6 +2312,7 @@ describe("apply release plan", () => {
           null
         ]
       },
+      undefined,
       setupFunc
     );
 
@@ -2294,6 +2395,7 @@ describe("apply release plan", () => {
             null
           ]
         },
+        undefined,
         setupFunc
       );