+++ packages/assemble-release-plan/src/index.ts
@@ -16,6 +16,11 @@ import { Packages, Package } from "@manypkg/get-packages";
 import { getDependentsGraph } from "@changesets/get-dependents-graph";
 import { PreInfo, InternalRelease } from "./types";
 
+type SnapshotReleaseParameters = {
+  tag?: string | undefined;
+  commit?: string | undefined;
+};
+
 function getPreVersion(version: string) {
   let parsed = semver.parse(version)!;
   let preVersion =
@@ -27,31 +32,64 @@ function getPreVersion(version: string) {
   return preVersion;
 }
 
-function getSnapshotSuffix(snapshot?: string | boolean): string | undefined {
-  if (snapshot === undefined) {
-    return;
+function getSnapshotSuffix(
+  template: Config["snapshot"]["prereleaseTemplate"],
+  snapshotParameters: SnapshotReleaseParameters
+): string {
+  let snapshotRefDate = new Date();
+
+  const placeholderValues = {
+    commit: snapshotParameters.commit,
+    tag: snapshotParameters.tag,
+    timestamp: snapshotRefDate.getTime().toString(),
+    datetime: snapshotRefDate
+      .toISOString()
+      .replace(/\.\d{3}Z$/, "")
+      .replace(/[^\d]/g, "")
+  };
+
+  // We need a special handling because we need to handle a case where `--snapshot` is used without any template,
+  // and the resulting version needs to be composed without a tag.
+  if (!template) {
+    return [placeholderValues.tag, placeholderValues.datetime]
+      .filter(Boolean)
+      .join("-");
   }
 
-  let dateAndTime = new Date()
-    .toISOString()
-    .replace(/\.\d{3}Z$/, "")
-    .replace(/[^\d]/g, "");
-  let tag = "";
+  const placeholders = Object.keys(placeholderValues) as Array<
+    keyof typeof placeholderValues
+  >;
 
-  if (typeof snapshot === "string") tag = `-${snapshot}`;
+  if (!template.includes(`{tag}`) && placeholderValues.tag !== undefined) {
+    throw new Error(
+      `Failed to compose snapshot version: "{tag}" placeholder is missing, but the snapshot parameter is defined (value: '${placeholderValues.tag}')`
+    );
+  }
 
-  return `${tag}-${dateAndTime}`;
+  return placeholders.reduce((prev, key) => {
+    return prev.replace(new RegExp(`\\{${key}\\}`, "g"), () => {
+      const value = placeholderValues[key];
+      if (value === undefined) {
+        throw new Error(
+          `Failed to compose snapshot version: "{${key}}" placeholder is used without having a value defined!`
+        );
+      }
+
+      return value;
+    });
+  }, template);
 }
 
-function getNewVersion(
+function getSnapshotVersion(
   release: InternalRelease,
   preInfo: PreInfo | undefined,
-  snapshotSuffix: string | undefined,
-  useCalculatedVersionForSnapshots: boolean
+  useCalculatedVersion: boolean,
+  snapshotSuffix: string
 ): string {
   if (release.type === "none") {
     return release.oldVersion;
   }
+
   /**
    * Using version as 0.0.0 so that it does not hinder with other version release
    * For example;
@@ -59,19 +97,24 @@ function getNewVersion(
    * and a consumer is using the range ^1.0.0-beta, most people would expect that range to resolve to 1.0.0-beta.0
    * but it'll actually resolve to 1.0.0-canary-hash. Using 0.0.0 solves this problem because it won't conflict with other versions.
    *
-   * You can set `useCalculatedVersionForSnapshots` flag to true to use calculated versions if you don't care about the above problem.
+   * You can set `snapshot.useCalculatedVersion` flag to true to use calculated versions if you don't care about the above problem.
    */
-  if (snapshotSuffix && !useCalculatedVersionForSnapshots) {
-    return `0.0.0${snapshotSuffix}`;
+  const baseVersion = useCalculatedVersion
+    ? incrementVersion(release, preInfo)
+    : `0.0.0`;
+
+  return `${baseVersion}-${snapshotSuffix}`;
+}
+
+function getNewVersion(
+  release: InternalRelease,
+  preInfo: PreInfo | undefined
+): string {
+  if (release.type === "none") {
+    return release.oldVersion;
   }
 
-  const calculatedVersion = incrementVersion(release, preInfo);
-
-  if (snapshotSuffix && useCalculatedVersionForSnapshots) {
-    return `${calculatedVersion}${snapshotSuffix}`;
-  }
-
-  return calculatedVersion;
+  return incrementVersion(release, preInfo);
 }
 
 function assembleReleasePlan(
@@ -80,7 +123,10 @@ function assembleReleasePlan(
   config: Config,
   // intentionally not using an optional parameter here so the result of `readPreState` has to be passed in here
   preState: PreState | undefined,
-  snapshot?: string | boolean
+  // snapshot: undefined            ->  not using snaphot
+  // snapshot: { tag: undefined }   ->  --snapshot (empty tag)
+  // snapshot: { tag: "canary" }    ->  --snapshot canary
+  snapshot?: SnapshotReleaseParameters
 ): ReleasePlan {
   let packagesByName = new Map(
     packages.packages.map(x => [x.packageJson.name, x])
@@ -93,9 +139,6 @@ function assembleReleasePlan(
   );
 
   const preInfo = getPreInfo(changesets, packagesByName, config, preState);
-
-  // Caching the snapshot version here and use this if it is snapshot release
-  const snapshotSuffix = getSnapshotSuffix(snapshot);
 
   // releases is, at this point a list of all packages we are going to releases,
   // flattened down to one release per package, having a reference back to their
@@ -158,18 +201,23 @@ function assembleReleasePlan(
     }
   }
 
+  // Caching the snapshot version here and use this if it is snapshot release
+  const snapshotSuffix =
+    snapshot && getSnapshotSuffix(config.snapshot.prereleaseTemplate, snapshot);
+
   return {
     changesets: relevantChangesets,
     releases: [...releases.values()].map(incompleteRelease => {
       return {
         ...incompleteRelease,
-        newVersion: getNewVersion(
-          incompleteRelease,
-          preInfo,
-          snapshotSuffix,
-          config.___experimentalUnsafeOptions_WILL_CHANGE_IN_PATCH
-            .useCalculatedVersionForSnapshots
-        )
+        newVersion: snapshotSuffix
+          ? getSnapshotVersion(
+              incompleteRelease,
+              preInfo,
+              config.snapshot.useCalculatedVersion,
+              snapshotSuffix
+            )
+          : getNewVersion(incompleteRelease, preInfo)
       };
     }),
     preState: preInfo?.state