+++ packages/cli/src/commands/add/index.ts
@@ -23,7 +23,9 @@ export default async function add(
   { empty, open }: { empty?: boolean; open?: boolean },
   config: Config
 ) {
-  const packages = await getPackages(cwd);
+  const packages = (await getPackages(cwd)).packages.filter(
+    pkg => !config.ignore.includes(pkg.packageJson.name)
+  );
   const changesetBase = path.resolve(cwd, ".changeset");
 
   let newChangeset: UnwrapPromise<ReturnType<typeof createChangeset>>;
@@ -39,10 +41,11 @@ export default async function add(
       ref: config.baseBranch
     });
     const changePackagesName = changedPackages
-      .filter(a => a)
-      .map(pkg => pkg.packageJson.name);
-    newChangeset = await createChangeset(changePackagesName, packages.packages);
-    printConfirmationMessage(newChangeset, packages.packages.length > 1);
+      .map(pkg => pkg.packageJson.name)
+      .filter(pkgName => !config.ignore.includes(pkgName));
+
+    newChangeset = await createChangeset(changePackagesName, packages);
+    printConfirmationMessage(newChangeset, packages.length > 1);
 
     if (!newChangeset.confirmed) {
       newChangeset = {