+++ packages/changelog-git/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @changesets/changelog-git
 
+## 0.1.12
+
+### Patch Changes
+
+- Updated dependencies [[`dd9b76f`](https://github.com/changesets/changesets/commit/dd9b76f162a546ae8b412e0cb10277f971f3585e)]:
+  - @changesets/types@5.1.0
+
 ## 0.1.11
 
 ### Patch Changes